+++ routes/api/contacts.js
@@ -4,21 +4,22 @@ const router = express.Router();
 
 const {
   getAllContacts,
-  // getById,
+  getById,
   addNewContact,
   // deleteContact,
   // changeContact,
 } = require("../../controllers/contacts");
 
-// const { validateBody } = require("../../middlewares");
+const { validateBody } = require("../../middlewares");
 // const schemas = require("../../schemas/contacts");
+const { schemas } = require("../../models/contact");
 // -----------------
 router.get("/", getAllContacts);
 
-// router.get("/:contactId", getById);
+router.get("/:contactId", getById);
 
-// router.post("/", validateBody(schemas.addSchemaPost), addNewContact);
-router.post("/", addNewContact);
+router.post("/", validateBody(schemas.addSchemaPost), addNewContact);
+// router.post("/", addNewContact);
 // router.delete("/:contactId", deleteContact);
 
 // router.put("/:contactId", validateBody(schemas.addSchemaPut), changeContact);
